% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustEval.R
\name{clustEval}
\alias{clustEval}
\title{Cluster-wise Clustering Robustness Evaluation}
\usage{
clustEval(dset, iteration, clust)
}
\arguments{
\item{dset}{(numeric matrix) features in rows and observations in columns}

\item{iteration}{(positive integer) specifies current iteration}

\item{clust}{return value of coreClust}
}
\value{
a numeric vector, specifies the clustering robustness (higher value
means more robust) of each clustering scheme
}
\description{
A sample cluster-wise clustering robustness evaluation framework (described
in "Examples" section, used as default in iterClust framework). Customized
frameworks can be defined following rules specified in "Usage", "Arguments"
and "Value" sections.
}
\examples{
clustEval <- function(dset, iteration, clust){
    dist <- as.dist(1 - cor(dset))
    clustEval <- vector("numeric", length(clust))
    for (i in 1:length(clust)){
        clustEval[i] <- mean(silhouette(clust[[i]], dist)[, "sil_width"])}
    return(clustEval)}

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{clustEval}
