% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustHetero.R
\name{clustHetero}
\alias{clustHetero}
\title{Cluster Heterogeneity Evaluation}
\usage{
clustHetero(clustEval, iteration)
}
\arguments{
\item{clustEval, }{return value of clustEval}

\item{iteration}{(positive integer) specifies current iteration}
}
\value{
a boolean vector, specifies whether a cluster is heterogenous
}
\description{
A sample cluster heterogeneity evaluation framework (described in "Examples"
section, used as default in iterClust framework). Customized frameworks can
be defined following rules specified in "Usage", "Arguments" and "Value"
sections.
}
\examples{
clustHetero <- function(clustEval, iteration){
    return(clustEval > 0*iteration+0.15)}

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{clustHetero}
