% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreClust.R
\name{coreClust}
\alias{coreClust}
\title{Clustering}
\usage{
coreClust(dset, iteration)
}
\arguments{
\item{dset}{(numeric matrix) features in rows and observations in columns}

\item{iteration}{(positive integer) specifies current iteration}
}
\value{
a list, each element contains clustering vectors (named numeric
vector with observation names as name and corresponding cluster number as
element) under a specific clustering parameter
}
\description{
A sample clustering framework (described in "Examples" section, used as
default in iterClust framework). Customized frameworks can be defined
following rules specified in "Usage", "Arguments" and "Value" sections.
}
\examples{
coreClust <- function(dset, iteration){
    dist <- as.dist(1 - cor(dset))
    range=seq(2, ncol(dset)-1, by = 1)
    clust <- vector("list", length(range))
    for (i in 1:length(range)) clust[[i]] <- pam(dist, range[i])$clustering
    return(clust)}

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{coreClust}
