% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelect.R
\name{featureSelect}
\alias{featureSelect}
\title{Feature Selection}
\usage{
featureSelect(dset, iteration, feature)
}
\arguments{
\item{dset}{(numeric matrix) features in rows and observations in columns}

\item{iteration}{(positive integer) specifies current iteration}

\item{feature}{(character array) specifies user defined features,
facilitating feature selection}
}
\value{
a character array, contains features selected
}
\description{
A sample feature selection framework (described in "Examples" section,
used as default in iterClust framework). Customized frameworks can be defined
following rules specified in "Usage", "Arguments" and "Value" sections.
}
\examples{
featureSelect <- function(dset, iteration, feature) return(rownames(dset))

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{featureSelect}
