% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsEval.R
\name{obsEval}
\alias{obsEval}
\title{Observation-wise Clustering Robustness Evaluation}
\usage{
obsEval(dset, clust, iteration)
}
\arguments{
\item{dset}{(numeric matrix) features in rows and observations in columns}

\item{clust}{optimal return value of coreClust}

\item{iteration}{(positive integer) specifies current iteration}
}
\value{
a numeric vector, specifies the clustering robustness (higher value
means more robust) of each observation under the optimal clustering scheme
}
\description{
A sample observation-wise clustering robustness evaluation framework
(described in "Examples" section, used as default in iterClust framework).
Customized frameworks can be defined following rules specified in "Usage",
"Arguments" and "Value" sections.
}
\examples{
obsEval <- function(dset, clust, iteration){
    dist <- as.dist(1 - cor(dset))
    obsEval <- vector("numeric", length(clust))
    return(silhouette(clust, dist)[, "sil_width"])}

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{obsEval}
