% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsOutlier.R
\name{obsOutlier}
\alias{obsOutlier}
\title{Outlier Observation Evaluation}
\usage{
obsOutlier(obsEval, iteration)
}
\arguments{
\item{obsEval, }{return value of obsEval}

\item{iteration}{(positive integer) specifies current iteration}
}
\value{
a boolean vector, specifies whether an observation is outlier
}
\description{
A sample outlier observation evaluation framework (described in "Examples"
section, used as default in iterClust framework). Customized frameworks can
be defined following rules specified in "Usage", "Arguments" and "Value"
sections.
}
\examples{
obsOutlier <- function(obsEval, iteration) return(obsEval < 0*iteration-1)

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{obsOutlier}
